# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Dokumentasi API v1.0.0",
        "contact": {
            "email": "prasdimasetyo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/skill-sets/": {
            "get": {
                "tags": [
                    "SkillSets"
                ],
                "summary": "Get all skill sets",
                "operationId": "c8e809d787d33adc98ca08ac1fb761fc",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": [
                                    {
                                        "candidate_id": 1,
                                        "skill_id": 2,
                                        "created_by": "2024-02-20T17:20:11.000000Z",
                                        "created_at": null,
                                        "updated_by": "2024-02-20T17:20:11.000000Z",
                                        "updated_at": null,
                                        "deleted_by": "2024-02-20T17:20:11.000000Z",
                                        "deleted_at": null
                                    },
                                    {
                                        "candidate_id": 2,
                                        "skill_id": 1,
                                        "created_by": null,
                                        "created_at": "2024-02-20T18:31:17.000000Z",
                                        "updated_by": null,
                                        "updated_at": "2024-02-20T18:31:17.000000Z",
                                        "deleted_by": null,
                                        "deleted_at": null
                                    },
                                    {
                                        "candidate_id": 2,
                                        "skill_id": 2,
                                        "created_by": null,
                                        "created_at": "2024-02-20T18:31:17.000000Z",
                                        "updated_by": null,
                                        "updated_at": "2024-02-20T18:31:17.000000Z",
                                        "deleted_by": null,
                                        "deleted_at": null
                                    },
                                    {
                                        "candidate_id": 3,
                                        "skill_id": 1,
                                        "created_by": null,
                                        "created_at": "2024-02-21T00:23:05.000000Z",
                                        "updated_by": null,
                                        "updated_at": "2024-02-21T00:23:05.000000Z",
                                        "deleted_by": null,
                                        "deleted_at": null
                                    },
                                    {
                                        "candidate_id": 3,
                                        "skill_id": 2,
                                        "created_by": null,
                                        "created_at": "2024-02-21T00:23:05.000000Z",
                                        "updated_by": null,
                                        "updated_at": "2024-02-21T00:23:05.000000Z",
                                        "deleted_by": null,
                                        "deleted_at": null
                                    }
                                ]
                            }
                        }
                    },
                    "404": {
                        "description": "SkillSet not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "SkillSet not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/skill-sets/{candidateId}/{skillId}": {
            "get": {
                "tags": [
                    "SkillSets"
                ],
                "summary": "Get a skill set by candidate ID and skill ID",
                "operationId": "2ee4c03eb40ae8276cb55325a9cac94c",
                "parameters": [
                    {
                        "name": "candidateId",
                        "in": "path",
                        "description": "Candidate ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "skillId",
                        "in": "path",
                        "description": "Skill ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "candidate_id": 1,
                                        "skill_id": 2,
                                        "created_by": "2024-02-20T17:20:11.000000Z",
                                        "created_at": null,
                                        "updated_by": "2024-02-20T17:20:11.000000Z",
                                        "updated_at": null,
                                        "deleted_by": "2024-02-20T17:20:11.000000Z",
                                        "deleted_at": null
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "SkillSet not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "SkillSet not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}